Tried to implement TextSlider.
