+++ src/ch/imedias/rsccfx/view/PopOverHelper.java
@@ -3,6 +3,7 @@ package ch.imedias.rsccfx.view;
 import ch.imedias.rsccfx.RsccApp;
 import ch.imedias.rsccfx.ViewController;
 import ch.imedias.rsccfx.model.Rscc;
+import ch.imedias.rsccfx.view.util.TextSlider;
 import javafx.beans.property.SimpleBooleanProperty;
 import javafx.geometry.Insets;
 import javafx.geometry.Rectangle2D;
@@ -18,10 +19,8 @@ import javafx.stage.Screen;
 import org.controlsfx.control.PopOver;
 import org.controlsfx.control.ToggleSwitch;
 
-
 /**
- * @author Lukas Marchesi
- * @date 17.04.2017.
+ * Shows popover for settings and help buttons.
  */
 public class PopOverHelper {
   // Get Screensize
@@ -41,8 +40,6 @@ public class PopOverHelper {
   private final double sliderWidth = overlayWidth / 1.2;
   private final double startXSlider = (overlayWidth / 2) - (sliderWidth / 2);
 
-  RsccApp rsccApp = new RsccApp();
-
   private SimpleBooleanProperty switchedOn = new SimpleBooleanProperty(false);
 
   private final Rscc model;
@@ -60,9 +57,6 @@ public class PopOverHelper {
   PopOver settingsPopOver = new PopOver();
   PopOver helpPopOver = new PopOver();
 
-  Text compressionSliderTxt = new Text();
-  Text qualitySliderTxt = new Text();
-
   Label requestCompressionLbl = new Label();
   Label requestQualityLbl = new Label();
   Label requestBgr233Lbl = new Label();
@@ -71,9 +65,10 @@ public class PopOverHelper {
   Label homeHelpLbl = new Label();
   Label requestHelpLbl = new Label();
   Label supporterHelpLbl = new Label();
-  Slider compressionSldr;
 
-  Slider qualitySldr;
+  TextSlider compressionSldr;
+
+  TextSlider qualitySldr;
 
   Pane compressionSliderPane = new Pane();
   Pane requestSettingsPane = new Pane();
@@ -114,56 +109,18 @@ public class PopOverHelper {
     // Settings PopOver TODO: StringsClass
 
     // Compression Settings
-    compressionSldr = new Slider(COMPRESSION_MIN, COMPRESSION_MAX, COMPRESSION_VALUE) {
-      @Override
-      protected void layoutChildren() {
-        super.layoutChildren();
+    compressionSldr = new TextSlider(COMPRESSION_MIN, COMPRESSION_MAX, COMPRESSION_VALUE);
 
-        Region thumb = (Region) lookup(".thumb");
-        if (thumb != null) {
-          compressionSliderTxt.setLayoutX(
-              thumb.getLayoutX()
-                  + thumb.getWidth() / 2
-                  - compressionSliderTxt.getLayoutBounds().getWidth() / 2
-                  + startXSlider
-          );
-        }
-      }
-    };
-
-    compressionSldr.setLayoutY(40);
     compressionSldr.setPrefWidth(sliderWidth);
     compressionSldr.setShowTickLabels(true);
     compressionSldr.setShowTickMarks(true);
-
-    compressionSliderTxt.setTextOrigin(VPos.TOP);
-    compressionSliderTxt.textProperty().bind(
-        compressionSldr.valueProperty().asString("%,.0f"));
-    compressionSliderTxt.getStyleClass().add("sliderTxts");
-    qualitySliderTxt.getStyleClass().add("sliderTxts");
 
     requestCompressionLbl.textProperty().set("Kompression");
     requestCompressionLbl.getStyleClass().add("sliderLbls");
 
     // Quality Settings
-    qualitySldr = new Slider(QUALITY_MIN, QUALITY_MAX, QUALITY_VALUE) {
-      @Override
-      protected void layoutChildren() {
-        super.layoutChildren();
+    qualitySldr = new TextSlider(QUALITY_MIN, QUALITY_MAX, QUALITY_VALUE);
 
-        Region thumb = (Region) lookup(".thumb");
-        if (thumb != null) {
-          qualitySliderTxt.setLayoutX(
-              thumb.getLayoutX()
-                  + thumb.getWidth() / 2
-                  - qualitySliderTxt.getLayoutBounds().getWidth() / 2
-                  + startXSlider
-          );
-        }
-      }
-    };
-
-    qualitySldr.setLayoutY(40);
     qualitySldr.setPrefWidth(sliderWidth);
     qualitySldr.setShowTickLabels(true);
     qualitySldr.setShowTickMarks(true);
@@ -172,8 +129,6 @@ public class PopOverHelper {
     requestQualityLbl.textProperty().set("Qualität");
     requestQualityLbl.getStyleClass().add("sliderLbls");
 
-    qualitySliderTxt.setTextOrigin(VPos.TOP);
-    qualitySliderTxt.textProperty().bind(qualitySldr.valueProperty().asString("%,.0f"));
 
     requestBgr233Lbl.textProperty().set("bgr233");
     requestBgr233Lbl.setId("requestBgr233Lbl");
@@ -184,12 +139,11 @@ public class PopOverHelper {
     requestBitCurrentSettingsLbl.textProperty().set("Ihre momentane Einstellung ist");
     requestBitCurrentSettingsLbl.setId("requestBitCurrentSettingsLbl");
 
-    compressionSliderPane.getChildren().addAll(compressionSldr, compressionSliderTxt);
-    qualitySliderPane.getChildren().addAll(qualitySldr, qualitySliderTxt);
+    qualitySliderPane.getChildren().addAll(qualitySldr);
 
     supportSettingsBox.setPadding(new Insets(10));
 
-    supportSettingsBox.getChildren().add(new VBox(compressionSliderPane, requestCompressionLbl));
+    supportSettingsBox.getChildren().add(new VBox(compressionSldr, requestCompressionLbl));
     supportSettingsBox.getChildren().add(new VBox(qualitySliderPane, requestQualityLbl));
     supportSettingsBox.getChildren().add(new HBox(eightBitTgl, requestBgr233Lbl));
     supportSettingsBox.getChildren().add(new HBox(requestViewOnlyLbl));
@@ -199,7 +153,6 @@ public class PopOverHelper {
     compressionSldr.setLayoutX(startXSlider);
     supportSettingsBox.setPrefWidth(overlayWidth);
     supportSettingsBox.setPrefHeight(overlayHeight);
-
 
     requestSettingsPane.getChildren().add(requestSettingsBox);
